requirements.txt
fastapi
uvicorn[standard]
pydantic
wikipedia
PyJWT
python-dotenv
gTTS
requests
